Crate calendar3

Features

Handle the following Resources with ease from the central hub ...

acl (delete, get, insert, list, patch, update and watch)

calendar list (delete, get, insert, list, patch, update and watch)

calendars (clear, delete, get, insert, patch and update)

channels (stop)

colors (get)

events (delete, get, import, insert, instances, list, move, patch, quickadd, update and watch)

freebusy (query)

settings (get, list and watch)

Everything else about the calendar API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.events().delete(...).doit()
let r = hub.events().insert(...).doit()
let r = hub.events().instances(...).doit()
let r = hub.events().quick_add(...).doit()
let r = hub.events().patch(...).doit()
let r = hub.events().import(...).doit()
let r = hub.events().move_(...).doit()
let r = hub.events().update(...).doit()
let r = hub.events().watch(...).doit()
let r = hub.events().get(...).doit()
let r = hub.events().list(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate calendar3;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and `client_secret`, 
// among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about what's going on
// You probably want to bring in your own `TokenStorage` to persist tokens and retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CalendarHub::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Channel = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.events().watch(&req, "calendarId")
             .updated_min("ea")
             .time_zone("no")
             .time_min("justo")
             .time_max("justo")
             .sync_token("et")
             .single_events(true)
             .show_hidden_invitations(true)
             .show_deleted(false)
             .shared_extended_property("Lorem")
             .q("et")
             .private_extended_property("duo")
             .page_token("aliquyam")
             .order_by("sea")
             .max_results(-55)
             .max_attendees(-75)
             .i_cal_uid("erat")
             .always_include_email(false)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Acl
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

AclDeleteMethodBuilder
Deletes an access control rule.

AclGetMethodBuilder
Returns an access control rule.

AclInsertMethodBuilder
Creates an access control rule.

AclListMethodBuilder
Returns the rules in the access control list for the calendar.

AclMethodsBuilder
A builder providing access to all methods supported on acl resources. It is not used directly, but through the CalendarHub hub.

AclPatchMethodBuilder
Updates an access control rule. This method supports patch semantics.

AclRule
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch (request|response)

update (request|response)

get (response)

insert (request|response)

AclRuleScope
The scope of the rule.

This type is not used in any activity, and only used as part of another schema.

AclUpdateMethodBuilder
Updates an access control rule.

AclWatchMethodBuilder
Watch for changes to ACL resources.

Calendar
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

clear (none)

get (response)

update (request|response)

patch (request|response)

insert (request|response)

delete (none)

CalendarClearMethodBuilder
Clears a primary calendar. This operation deletes all events associated with the primary calendar of an account.

CalendarDeleteMethodBuilder
Deletes a secondary calendar. Use calendars.clear for clearing all events on primary calendars.

CalendarGetMethodBuilder
Returns metadata for a calendar.

CalendarHub
Central instance to access all CalendarHub related resource activities

CalendarInsertMethodBuilder
Creates a secondary calendar.

CalendarList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

CalendarListDeleteMethodBuilder
Deletes an entry on the user's calendar list.

CalendarListEntry
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

update (request|response)

patch (request|response)

insert (request|response)

CalendarListEntryNotificationSettings
The notifications that the authenticated user is receiving for this calendar.

This type is not used in any activity, and only used as part of another schema.

CalendarListGetMethodBuilder
Returns an entry on the user's calendar list.

CalendarListInsertMethodBuilder
Adds an entry to the user's calendar list.

CalendarListListMethodBuilder
Returns entries on the user's calendar list.

CalendarListMethodsBuilder
A builder providing access to all methods supported on calendarList resources. It is not used directly, but through the CalendarHub hub.

CalendarListPatchMethodBuilder
Updates an entry on the user's calendar list. This method supports patch semantics.

CalendarListUpdateMethodBuilder
Updates an entry on the user's calendar list.

CalendarListWatchMethodBuilder
Watch for changes to CalendarList resources.

CalendarMethodsBuilder
A builder providing access to all methods supported on calendar resources. It is not used directly, but through the CalendarHub hub.

CalendarNotification
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

CalendarPatchMethodBuilder
Updates metadata for a calendar. This method supports patch semantics.

CalendarUpdateMethodBuilder
Updates metadata for a calendar.

Channel
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

stop (request)

watch (request|response)

watch (request|response)

watch (request|response)

watch (request|response)

ChannelMethodsBuilder
A builder providing access to all methods supported on channel resources. It is not used directly, but through the CalendarHub hub.

ChannelStopMethodBuilder
Stop watching resources through this channel

ColorDefinition
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

ColorGetMethodBuilder
Returns the color definitions for calendars and events.

ColorMethodsBuilder
A builder providing access to all methods supported on color resources. It is not used directly, but through the CalendarHub hub.

Colors
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

Error
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Event
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete (none)

insert (request|response)

instances (none)

quickAdd (response)

patch (request|response)

import (request|response)

move (response)

update (request|response)

watch (none)

get (response)

list (none)

EventAttachment
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

EventAttendee
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

EventCreator
The creator of the event. Read-only.

This type is not used in any activity, and only used as part of another schema.

EventDateTime
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

EventDeleteMethodBuilder
Deletes an event.

EventExtendedProperties
Extended properties of the event.

This type is not used in any activity, and only used as part of another schema.

EventGadget
A gadget that extends this event.

This type is not used in any activity, and only used as part of another schema.

EventGetMethodBuilder
Returns an event.

EventImportMethodBuilder
Imports an event. This operation is used to add a private copy of an existing event to a calendar.

EventInsertMethodBuilder
Creates an event.

EventInstanceMethodBuilder
Returns instances of the specified recurring event.

EventListMethodBuilder
Returns events on the specified calendar.

EventMethodsBuilder
A builder providing access to all methods supported on event resources. It is not used directly, but through the CalendarHub hub.

EventMoveMethodBuilder
Moves an event to another calendar, i.e. changes an event's organizer.

EventOrganizer
The organizer of the event. If the organizer is also an attendee, this is indicated with a separate entry in attendees with the organizer field set to True. To change the organizer, use the move operation. Read-only, except when importing an event.

This type is not used in any activity, and only used as part of another schema.

EventPatchMethodBuilder
Updates an event. This method supports patch semantics.

EventQuickAddMethodBuilder
Creates an event based on a simple text string.

EventReminder
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

EventReminders
Information about the event's reminders for the authenticated user.

This type is not used in any activity, and only used as part of another schema.

EventSource
Source of an event from which it was created; for example a web page, an email message or any document identifiable by an URL using HTTP/HTTPS protocol. Accessible only by the creator of the event.

This type is not used in any activity, and only used as part of another schema.

EventUpdateMethodBuilder
Updates an event.

EventWatchMethodBuilder
Watch for changes to Events resources.

Events
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

instances (response)

list (response)

FreeBusyCalendar
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

FreeBusyGroup
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

FreeBusyRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

query (request)

FreeBusyRequestItem
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

FreeBusyResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

query (response)

FreebusyMethodsBuilder
A builder providing access to all methods supported on freebusy resources. It is not used directly, but through the CalendarHub hub.

FreebusyQueryMethodBuilder
Returns free/busy information for a set of calendars.

Setting
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (none)

watch (none)

get (response)

SettingGetMethodBuilder
Returns a single user setting.

SettingListMethodBuilder
Returns all user settings for the authenticated user.

SettingMethodsBuilder
A builder providing access to all methods supported on setting resources. It is not used directly, but through the CalendarHub hub.

SettingWatchMethodBuilder
Watch for changes to Settings resources.

Settings
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

TimePeriod
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.