Struct calendar3::EventMethodsBuilder

pub struct EventMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on event resources. It is not used directly, but through the CalendarHub hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate calendar3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use calendar3::CalendarHub;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CalendarHub::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `delete(...)`, `get(...)`, `import(...)`, `insert(...)`, `instances(...)`, `list(...)`, `move_(...)`, `patch(...)`, `quick_add(...)`, `update(...)` and `watch(...)`
// to build up your call.
let rb = hub.events();

Methods

impl<'a, C, NC, A> EventMethodsBuilder<'a, C, NC, A>

fn delete(&self, calendar_id: &str, event_id: &str) -> EventDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes an event.

fn insert(&self, request: &Event, calendar_id: &str) -> EventInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates an event.

fn import(&self, request: &Event, calendar_id: &str) -> EventImportMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Imports an event. This operation is used to add a private copy of an existing event to a calendar.

fn instances(&self, calendar_id: &str, event_id: &str) -> EventInstanceMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns instances of the specified recurring event.

fn get(&self, calendar_id: &str, event_id: &str) -> EventGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns an event.

fn list(&self, calendar_id: &str) -> EventListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns events on the specified calendar.

fn patch(&self, request: &Event, calendar_id: &str, event_id: &str) -> EventPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an event. This method supports patch semantics.

fn move_(&self, calendar_id: &str, event_id: &str, destination: &str) -> EventMoveMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Moves an event to another calendar, i.e. changes an event's organizer.

fn update(&self, request: &Event, calendar_id: &str, event_id: &str) -> EventUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an event.

fn watch(&self, request: &Channel, calendar_id: &str) -> EventWatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Watch for changes to Events resources.

fn quick_add(&self, calendar_id: &str, text: &str) -> EventQuickAddMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates an event based on a simple text string.